Set a description in the alert when updating (#7)
For some weird reasons the optional field Description is not always
optional. It seems that you can create an alert w/o description but you
can't edit it. This caused errors when the alert was enabled or disabled
so this commit adds a dummy description if the alert was created without
one.